Create standard routing structure for api server
To further break out the server express app logic, will use express's
Router class to create routers for each endpoint. The endpoints will be
registered with app.use within /api/index.js, and then the index is simply required within the express config file.

Services for querying external apis will now reside in a top level directory '/services'.
