+++ api/index.js
@@ -1,0 +1,3 @@
+module.exports = function(app) {
+  app.use('/search', require('./routes/search'));
+};
